+++ tradingstrategy/universe.py
@@ -53,3 +53,4 @@ class Universe:
         """
         assert len(self.exchanges) == 1
         return self.exchanges[0]
+